refactor: extract auto-track opt-in machinery into an autotrack submodule for r8
The auto-track additions pushed extension_bridge/mod.rs to 1425 LOC, over the
1400 hard cap. Move the opt-in file load/persist, the enabled accessors, and the
autotrack result reply into a sibling autotrack.rs (mirrors status_update.rs).
mod.rs is now 1372. No behavior change.
